// src/app/page/master/pegawai-sk-struktur-nomor/pegawai-struktur-nomor.interface.ts
import {EventEmitter} from '@angular/core';

export interface PegawaiStrukturNomor {
	kdProfile?: string;
	kdAgama?: string;
	namaAgama?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;

 }